requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

langgraph==0.2.35
langchain-core==0.3.21

# Optional: OpenAI-compatible client (works for OpenAI and many compatibles)
openai==1.55.3
httpx==0.27.2
